Restored Windows support

+++ src/sagotmx/tmx_struct.h
@@ -626,7 +626,7 @@ inline uint32_t getTileFromLayer(const TileMap& m, const TileLayer& l, int x, in
 	size_t tile_index = (m.height*y+x)*sizeof(uint32_t);
 	if (tile_index > l.data.payload.size()-sizeof(uint32_t) ) {
 		throw SagoTiledException("ERROR: getTileFromLayer called with coordinates out-of-bound. Called with (%d, %d). Limit (%d, %d). Or the layer is corrupt. "
-				"Reported number of tiles in layer: %ld", x, y, m.width-1, m.height-1, l.data.payload.size()/sizeof(uint32_t));
+				"Reported number of tiles in layer: %ld", x, y, m.width-1, m.height-1, (long int)l.data.payload.size()/sizeof(uint32_t));
 	}
 	const unsigned char *data = reinterpret_cast<const unsigned char*>(l.data.payload.data());
 	uint32_t global_tile_id = data[tile_index] |
@@ -641,7 +641,7 @@ inline void setTileOnLayerNumber(TileMap& m, int layer_number, int x, int y, uin
 	TileLayer& l = m.layers.at(layer_number);
 	if (tile_index > l.data.payload.size()-sizeof(uint32_t) ) {
 		throw SagoTiledException("ERROR: setTileOnLayerNumber called with coordinates out-of-bound. Called with (%d, %d). Limit (%d, %d). Or the layer is corrupt. "
-				"Reported number of tiles in layer: %ld", x, y, m.width-1, m.height-1, l.data.payload.size()/sizeof(uint32_t));
+				"Reported number of tiles in layer: %ld", x, y, m.width-1, m.height-1, (long int)l.data.payload.size()/sizeof(uint32_t));
 	}
 	std::string& data = l.data.payload;
 	uint32_t* dp = reinterpret_cast<uint32_t*> (&data[tile_index]);
